Backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
assemblyai==0.23.0
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
Markdown==3.5.2
matplotlib-inline==0.1.6
nltk==3.8.1
openai==1.13.3
packaging==24.0
parso==0.8.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
pytube==15.0.0
regex==2023.12.25
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
yt-dlp==2025.8.22
